Conditional run fixSyntheticEventsOnDisabled logic
can-util/dom/events/events contains a check for when disabled behaviors
needs to be adjusted. This feature detection takes place in the body of
module, and needs DOM access, which it cannot get in Node. Since this
logic is Firefox-specific, using isBrowserWindow to prevent this feature
detection from happening in Node. Closes #221